BMPTest

Description

This module is used as a test base for Haskell Image processing techniques

Documentation

doBoth

Arguments

:: String
Path to Input File (BMP)

-> String
Path to Output File (BMP)

-> IO ()

runs both loadBitmap and writeRaw

showBMPInfo :: Maybe BMP -> IO ()

This Function is currently not functional

makeComponents :: IO ()

writeRaw :: String -> Maybe BMP -> IO ()

Gets the red component from an array of Word8s

Test Of the Comments

loadBitmap :: String -> IO (Maybe BMP)

loadBitmap : Loads a 24 or 32 bit BMP. This obviously requires that the image is color