FIP

Description

This is the primary Module for Functional Image Processing Course | Developed by Andrew Kordik for | the University of Dayton | All Rights Reserved | This library is still highly volitile. | The goal is to provide an interface to | Discrete Signal Processing with a Functional Approach | by leveraging Haskell and demonstraiting it through | an image processing libarary, which means we are | using two dimentional signal spaces

Documentation

data Image i e

This is the core type of the library, each component of the image is stored as an Array of Word8s

Constructors

Image

Fields

width :: Int

Width in pixels of the image

height :: Int

Height in pixels of the image

red :: Array i e

The Red pixel Values of the image

green :: Array i e

The Green pixel values of the image

blue :: Array i e

The Blue pixel values of the image

alpha :: Array i e

The Alpha pixel values of the image

class ValueMappable a where

The Type-Class ValueMappable manipulates Arrays without changing thier dimensions. | There is no checking to ensure that this is maintained.

Methods

valueMap

Arguments

:: (Array (Int, Int) a -> Array (Int, Int) b)
A Function that can manipulate a 2-D Array

-> Image (Int, Int) a
The input Image

-> Image (Int, Int) b
The output Image

class IndexMappable a where

The Type-Class IndexMappable maniplates Arrays and may-or-maynot change their dimenions. | There is no checking to ensure that the dimension changes are the same as the changes to the array

Methods

indexMap

Arguments

:: (Int -> Int)
A Function that changes the width of the image

-> (Int -> Int)
A Function that changes the height of the image

-> (Array (Int, Int) a -> Array (Int, Int) b)
A Function that manipulates a 2-D Array

-> Image (Int, Int) a
The input Image

-> Image (Int, Int) b
The output Image

applyWindow

Arguments

:: (RealFrac a, Integral a1, Integral e)

=> Array (Int, Int) a

-> Array (Int, Int) a1

-> Array (Int, Int) e

applyWindow takes a 2-D array that contains the filter values. This | is synonmous with a Window filter common in Image Processing Algorithms. | The Window Filter is applied as a sum of products on an element my element | basis. First some basic information about the Array being processed (imageArray) | and the window being applied (window) are gathered (i.e. width and height). | Next the imageArray is padded with zeros. This is common but not required by | Windowing techniques. The padding is accomplished by extending the image Array | in all directions, such that the values are at the same index as the original, | In other words, that is the padding data actually exists at indicies below zero

loadImage

Arguments

:: String
Input filename of a Bitmap Image with extension

-> IO (Maybe (Image (Int, Int) Word8))
Loaded Image

loadImage takes a filename including the extension of a 32bit Bitmap image and returns an Image wrapped in a Maybe and IO monad If the load fails, An Error is Printed and the function returns IO (Nothing) Otherwise it returns IO( Just Image (Int,Int) Word8))

writeImage

Arguments

:: String
The Desired output filename

-> Image (Int, Int) Word8
The Image data to be written as a Bitmap

-> IO ()
Empty IO Monad

writeImage takes a file name and an image to be written as a 32-bit | Bitmap Image

getWidthAndHeight

Arguments

:: ((Int, Int), (Int, Int))
((wStart, hStart), (wEnd, hEnd))

-> (Int, Int)
(width, height)

Extracts width and height from a bounds representation

loadBitmap :: String -> IO (Maybe BMP)

loadBitmap : Loads a 24 or 32 bit BMP. This obviously requires that the image is stored in a Color Bitmap file.

bmpToImage :: BMP -> Image (Int, Int) Word8

This takes an a BMP type and coverts it to an Image type so that it can be worked with

imageToBmp :: Image (Int, Int) Word8 -> BMP

imageToBmp takes the custom image type and puts it into a 32bit RGBA Bitmap

arrayToByteString

Arguments

:: (Ix t2, Ix t1, Num t1, Num t2, Enum t1, Enum t2)

=> Array (t1, t2) t
a 2-D Array indexed from 0

-> t1
the width in pixels / array elements

-> t2
the height in pixels / array elements

-> [t]
a 1 dimentional list in raster order of all pixels/elements

arrayToByteString takes a 2-D array and moves it to a 1-D list so that it can be packed into a bytestring

combineComponents

Arguments

:: [a]
An array of Red components only

-> [a]
An array of Green components only

-> [a]
An array of Blue components only

-> [a]
An array of Alpha components only

-> [a]
An array of All components in raster order RGBA

combineComponents takes arrays containing a single component and aggregates them into a single multiple component array in Raster order (RGBA)

byteStringToArray

Arguments

:: ByteString
A Single compnent bytestring

-> Int
Width in Pixels

-> Int
Height in Pixels

-> Array (Int, Int) Word8
2-d Array containing a single component

byteStringToArray is takes a ByteString and converts it to a 2-d array of the appropriate width and height. Arrays are indexed from 0,0 to width-1, height-1

getOnlyRed

Arguments

:: [a]
A List containing RGBA data

-> [a]
A List containing only Red data

Puts only the red components into a standalone array

getOnlyGreen

Arguments

:: [a]
A List containing RGBA data

-> [a]
A List containing only Green data

Puts only the Green components into a standalone array

getOnlyBlue

Arguments

:: [a]
A List containing RGBA data

-> [a]
A List containing only Blue data

Puts only the Blue components into a standalone array

getOnlyAlpha

Arguments

:: [a]
A List containing RGBA data

-> [a]
A List containing only Alpha data

Puts only the Alpha components into a standalone array